refactor(stt): extract common UI and backend logic for STT providers
